backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django_filter==22.1
django-simple-captcha==0.5.17
django-auditlog==2.1.1
mysqlclient==2.1.1
gunicorn==20.1.0
celery
minio
# 新增
django-cors-headers==3.6.0
drf-nested-routers
django-prometheus==1.0.4
django-cors-headers==3.6.0
django-soft-delete==0.9.21
drf-spectacular==0.22.0
drf-spectacular-sidecar
djangorestframework-simplejwt
# packages
requests
openpyxl
pandas
chardet